requirements.txt
Django==2.2
flake8==3.8.3
psycopg2==2.8.5
python-decouple==3.3
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.7
python-dotenv==0.14.0
django-easy-select2==1.5.7







